refactor: 분류 목록 조회 - API호출 위치 변경
